test(shell): reorganize Playwright E2E tests into feature-focused suites
Redistributes E2E tests from phase-based and scattered files into
cohesive, topic-based test suites:
- workspace-routing-bootstrapping.e2e.ts (app boot, URL params, layout
  collapse)
- settings-and-configuration.e2e.ts (persisting configs, 1P/3P env,
  config locking)
- renderer-integration-and-telemetry.e2e.ts (preview handshakes, event
  sync, overlay blockers)
- debug-drawer-and-logs.e2e.ts (collapsible drawer, console logs, error
  badges, session resets)
- user-journey.e2e.ts (comprehensive workspace onboarding journey)

Cleans up and removes all deprecated phase-based and individual gap
files.

Use a window-level sentinel instead of page.waitForNavigation() to
robustly wait for location.reload() to complete, avoiding context
destroyed errors.
