+++ shell/e2e/debug-drawer-and-logs.e2e.ts
@@ -40,22 +40,18 @@ test.beforeEach(async ({page}) => {
   await page.goto('/?renderer=http://custom-renderer.com/index.html');
 });
 
-test.describe('Phase 6 Interactive Debugging Panels', () => {
-  test('verifies collapsible debug drawer and global clear sweeps', async ({page}) => {
-    // 1. Verify debug panel is expanded by default on mount in 1P mode
+test.describe('Debugging Panels & Diagnostic Logs', () => {
+  test('verifies collapsible debug drawer layout and global clear sweeps', async ({page}) => {
     const debugSection = page.locator('.debug-section');
     await expect(debugSection).not.toHaveClass(/collapsed/);
 
-    // 2. Collapse debug drawer via toggle button
     const collapseBtn = page.getByRole('button', {name: 'Toggle Debug Panel'});
     await collapseBtn.click();
     await expect(debugSection).toHaveClass(/collapsed/);
 
-    // 3. Expand debug drawer back
     await collapseBtn.click();
     await expect(debugSection).not.toHaveClass(/collapsed/);
 
-    // 4. Locate iframe and evaluate postMessage events originating from iframe
     const iframeBody = page.frameLocator('iframe.preview-iframe').locator('body');
     await expect(iframeBody).toBeVisible();
     await page.waitForTimeout(1000);
@@ -90,65 +86,79 @@ test.describe('Phase 6 Interactive Debugging Panels', () => {
       window.parent.postMessage(msg, '*');
     }, errorMsg);
 
-    // 5. Switch to Events tab to verify it mapped the event correctly
+    // Verify Events tab mapping
     await page.getByRole('tab', {name: 'Events'}).click();
     const eventRow = page.locator('.events-container table tr.element-row');
     await expect(eventRow).toHaveCount(1);
     await expect(page.locator('.events-container table')).toContainText('e2e_button_click');
 
-    // 6. Switch to Errors tab to verify exception consolidation and expansion
+    // Verify Errors tab consolidation & stack trace expansion
     await page.getByRole('tab', {name: 'Errors'}).click();
     const errorRow = page.locator('.errors-container table tr.element-row');
     await expect(errorRow).toHaveCount(1);
     await expect(page.locator('.errors-container table')).toContainText('E2E Exception trace');
 
-    // Expand the stack trace detail panel
     const expandBtn = page.getByRole('button', {name: 'Toggle Stack Trace'});
     await expandBtn.click();
     await expect(page.locator('.stack-preview')).toBeVisible();
     await expect(page.locator('.stack-preview')).toContainText('Error: E2E Failure');
 
-    // 7. Switch to Data Model tab and enter valid JSON
+    // Verify Data Model tab valid/invalid JSON alerts
     await page.getByRole('tab', {name: 'Data Model'}).click();
     const dataModelField = page.locator('.data-model-field textarea');
     await dataModelField.fill('{"valid": true}');
     await expect(page.locator('.invalid-json-badge')).not.toBeVisible();
 
-    // Enter invalid JSON to assert syntax warning badge appears
     await dataModelField.fill('{"invalid": }');
     await expect(page.locator('.invalid-json-badge')).toBeVisible();
 
-    // 8. Click Global Clear Sweeps button
+    // Verify Global Clear
     const clearBtn = page.getByRole('button', {name: 'Clear Logs'});
     await clearBtn.click();
 
-    // Assert Data Model syntax warning remains (not a log, should be unaffected)
+    // Warnings remain, but logs clear
     await expect(page.locator('.invalid-json-badge')).toBeVisible();
 
-    // Assert Events table is empty (shows placeholder)
     await page.getByRole('tab', {name: 'Events'}).click();
     await expect(page.locator('.events-placeholder')).toBeVisible();
 
-    // Assert Errors table is empty (shows placeholder)
     await page.getByRole('tab', {name: 'Errors'}).click();
     await expect(page.locator('.errors-placeholder')).toBeVisible();
   });
 
+  test('verifies console log warning routing from preview frame to Errors tab', async ({page}) => {
+    const iframeBody = page.frameLocator('iframe.preview-iframe').locator('body');
+    await expect(iframeBody).toBeVisible();
+    await page.waitForTimeout(1000);
+
+    const logMsg = {
+      type: PreviewBridgeMessageType.CONSOLE_LOG,
+      payload: {level: 'warn', message: 'Telemetry active warn'},
+    };
+    await iframeBody.evaluate((_, msg) => {
+      window.parent.postMessage(msg, '*');
+    }, logMsg);
+
+    await page.getByRole('tab', {name: 'Errors'}).click();
+    const errorRow = page.locator('.errors-container table tr.element-row').first();
+    await expect(errorRow).toBeVisible();
+    const rowText = await errorRow.textContent();
+    expect(rowText).toContain('warn');
+    expect(rowText).toContain('console');
+    expect(rowText).toContain('Telemetry active warn');
+  });
+
   test('verifies unread tab badging animations and resets', async ({page}) => {
-    // Start in Data Model tab (index 0)
     await page.getByRole('tab', {name: 'Data Model'}).click();
 
-    // Events badge should be hidden initially
     const eventsTabLabel = page.locator('.mat-mdc-tab-label-container .mdc-tab:nth-child(2)');
     const eventsBadge = eventsTabLabel.locator('.mat-badge-content');
     await expect(eventsBadge).toBeHidden();
 
-    // Errors badge should be hidden initially
     const errorsTabLabel = page.locator('.mat-mdc-tab-label-container .mdc-tab:nth-child(3)');
     const errorsBadge = errorsTabLabel.locator('.mat-badge-content');
     await expect(errorsBadge).toBeHidden();
 
-    // Locate iframe and evaluate postMessage unread events originating from iframe
     const iframeBody = page.frameLocator('iframe.preview-iframe').locator('body');
     await expect(iframeBody).toBeVisible();
     await page.waitForTimeout(1000);
@@ -180,18 +190,41 @@ test.describe('Phase 6 Interactive Debugging Panels', () => {
       window.parent.postMessage(msg, '*');
     }, crashMsg);
 
-    // Assert unread badges incremented and appear visible
     await expect(eventsBadge).toBeVisible();
     await expect(eventsBadge).toContainText('1');
     await expect(errorsBadge).toBeVisible();
     await expect(errorsBadge).toContainText('1');
 
-    // Switch to Events Tab and verify it resets events unread counter to 0 and hides
     await page.getByRole('tab', {name: /Events/}).click();
     await expect(eventsBadge).toBeHidden();
 
-    // Switch to Errors Tab and verify it resets errors unread counter to 0 and hides
     await page.getByRole('tab', {name: /Errors/}).click();
     await expect(errorsBadge).toBeHidden();
+  });
+
+  test('verifies New Session reset button clears localStorage session cache', async ({page}) => {
+    await page.evaluate(() => {
+      try {
+        localStorage.setItem('a2ui_composer_session_state', 'test_value');
+      } catch (e) {}
+    });
+
+    await page.evaluate(() => {
+      (window as any).__BEFORE_RELOAD__ = true;
+    });
+
+    await page.getByRole('button', {name: 'New Session'}).click();
+
+    // Wait for the page reload navigation to complete
+    await page.waitForFunction(() => (window as any).__BEFORE_RELOAD__ === undefined);
+
+    const testVal = await page.evaluate(() => {
+      try {
+        return localStorage.getItem('a2ui_composer_session_state');
+      } catch (e) {
+        return null;
+      }
+    });
+    expect(testVal).toBeNull();
   });
 });
